Added Suffix/Prefix Change for Bold & Fixed how Markdown is Applied
Added
Users can now modify the Prefix/Suffix of the Markdown. (FOR BOLD ONLY)
Users can now check if a specific Markdown is enabled
adding \ in front of the markdown will cancel it out
ex. \**bold** will send **bold** in stead of bold.

Changed
If your message between the Prefix and the Suffix is empty, it will not delete the Prefix/Suffix.
Before if you typed ** without anything else, it would delete the ** because the code assumed that it was applying Italic Styling.